# Improve query analysis