## The performance of query decomposition between main branch and this pr

dataset: 8 samples

**Main branch**

```
without fewshot examples
Average Metric: 4.5 / 8  (56.2): 100%|██████████| 8/8 [00:06<00:00,  1.17it/s] 
56.25

with fewshot examples
Average Metric: 7.275 / 8  (90.9): 100%|██████████| 8/8 [00:06<00:00,  1.31it/s]
90.94
```

**This PR**

```
without fewshot examples
Average Metric: 6.3 / 8  (78.8): 100%|██████████| 8/8 [00:04<00:00,  1.70it/s] 
78.75

with fewshot examples
Average Metric: 7.000000000000001 / 8  (87.5): 100%|██████████| 8/8 [00:05<00:00,  1.59it/s] 
87.5
```